% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{AssignStates}
\alias{AssignStates}
\title{Assign the polarization states in Immune Dictionary}
\usage{
AssignStates(
  object,
  state_list,
  include_other_cyt = FALSE,
  dist_thres = 0.8,
  marker_quantile = 0.9,
  embedding = "uce",
  center_unpolar = TRUE
)
}
\arguments{
\item{object}{A Seurat object of a cell type from the Immune Dictionary.}

\item{state_list}{A list of 4~6 elements of polarization states. Each element
should be a list of 2 variables: marker and cytokine. See the data(polar_states)
for a full list of polarization states.}

\item{include_other_cyt}{Logical. Whether to considering the non-driving cytokines
when assigning polarization states. If FALSE, only cells from samples treated
with driving cytokines may be assigned to the polarized states. If TRUE, all
cells treated by any cytokines may be assigned to the polarized states.
Default: FALSE.}

\item{dist_thres}{A numeric value in 0~1. Cells with cosine distance bigger
than this threshold will not be assigned to the polarized states.
Default: 0.2.}

\item{marker_quantile}{A numeric value in 0~1. Cells with the mean marker gene
expression smaller than this quantile will not be assigned into the polarized
states. Default: 0.9.}

\item{embedding}{The name of embedding saving the cell embeddings from the
single-cell foundation model in the Seurat object. Default: uce.}

\item{center_unpolar}{Logical. Whether to use unpolarized cell center as the
initial point for cell embedding vectors. The cosine distance is calculated
using the ell embedding vectors.If FALSE, the center of all cells is used as
the initial point for cell embedding vectors. Default: TRUE.}
}
\value{
A Seurat object with an additional variable 'polar' corresponding to
polarization states.
}
\description{
Assign the polarization states to a cell type in Immune Dictionary based on
Universial Cell Embeddings and marker gene expression.
}
\details{
Run this function only if you want to retrain the models using custom parameters!
}
\examples{
\dontrun{
# Please download the Immune Dictionary data at
cd8t <- AssignStates(cd8t, polar_states_cd8t, dist_thres=0.85)
}

}
