% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_polarization.R
\name{LoadPolarParams}
\alias{LoadPolarParams}
\title{Load the trained models and parameters}
\usage{
LoadPolarParams(celltype, verbose = TRUE)
}
\arguments{
\item{celltype}{A string of cell type whose polarization will be measured.
Must be one of: 'B', 'NK', 'CD8T', 'CD4T', 'Treg', 'Tgd', 'pDC',
'cDC1', 'cDC2', 'MigDC', 'LC', 'Macro', 'Mono', 'Neu'.}

\item{verbose}{Whether to message the loading progress.
Default: TRUE.}
}
\value{
A list containing 7 variables:
\describe{
\item{mean}{The mean of Universial Cell Embeddings of all cells in the training
dataset.}
\item{sd}{The standard deviation of Universial Cell Embeddings of all cells
in the training dataset.}
\item{mean_unpolar}{The mean of Universial Cell Embeddings of unpolarized cells
in the training dataset.}
\item{pc_loadings}{The principal component loadings to transform Universial
Cell Embeddings into principal component embeddings.}
\item{models}{The trained machine learning models to predict the polarization
scores based on principal component embeddings for each cell polarization
state.}
\item{unpolar_responses}{The polarization scores of unpolarized cells in the
training dataset. The probability of input data being unpolarized is calculated
by comparing to the distribution of these responses.}
\item{calibration_quantiles}{The 1~99 quantiles of nonconformity scores in
calibration data, used for conformal prediction.}
}
}
\description{
Load the trained models and parameters for measuring the polarization in a cell type.
}
\details{
This function will load a global variable named 'polar_params'
containing trained machine learning models and parameters.
}
\examples{
params_cd8t <- LoadPolarParams('CD8T')

}
