% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_polarization.R
\name{MeasurePolar}
\alias{MeasurePolar}
\title{Measure polarization of immune cells}
\usage{
MeasurePolar(
  object,
  celltype,
  embedding = "uce",
  unpolarized_cell = NA,
  return.df = FALSE,
  pc = 1:20,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object with an assay of UCE embedding for immune cell
polarization measurement. It is recommended that the object only contain one
cell type of interest.}

\item{celltype}{A string of cell type whose polarization will be measured.
Must be one of: B, NK, CD8T, CD4T, Treg, Tgd, pDC,
cDC1, cDC2, MigDC, LC, Macro, Mono, Neu.}

\item{embedding}{The name of dimension reduction or assay saving the cell
embeddings from the single-cell foundation model in the Seurat object.
Default: uce.}

\item{unpolarized_cell}{The name of cells that are considered unpolarized cells.
See details for selecting unpolarized cells. Default: NA.}

\item{return.df}{Whether to return the results as a dataframe. If TRUE,
return a dataframe. Else, return the Seurat object with updated meta.data.
Default: FALSE.}

\item{pc}{The index of principal component to be used by machine learning models.
Default: 1:20. DO NOT change this parameter if you uses the trained model provided
in the package! Only change it when you have trained custom models using
\code{CalculateParams}.}

\item{verbose}{Whether to message the progress in different polarization states.
Default: TRUE.}
}
\value{
Return either a dataframe or a Seurat object with updated meta.data,
depending on param return.df.
The dataframe includes 3 * N(polarization states) columns.
For each cell type, there are 4~6 polarization states.
For each polarization state, the polarization scores (range 0~1), P values,
and adjusted P values are calculated for all cells.
For details of polarization states, please refer to \code{data(polar_states)}
and the reference (Cui et al. Nature. 2024).
}
\description{
This is the main function to measure the polarization of immune cells based on
their Universial Cell Embeddings. The function uses trained machine learning
models to predict the polarization of cells in the input data.
}
\details{
Providing the unpolarized cell names is optional, but highly recommended.
The function will correct cell embedding batch effect by mapping the center
of unpolarized cells in the input dataset to the center of unpolarized cells
in the training dataset, thereby improving the measurement of polarization scores.
Here are some criteria for selecting unpolarized cells:
\enumerate{
\item When there are untreated samples and samples treated with certain stimulation
like cytokines or drugs, use the cells in untreated samples.
\item When there is a cell clusters with high expression of genes related to
steady state, like C1Q-high macrophages, use the cells in that cell cluster.
}

If not provided, the cell embedding batch effect between the training dataset
and the input dataset will not be corrected. Despite that, we found that the
method could generate results similar as the results when providing unpolarized
cells in several independent datasets, potentially benefiting from UCE's
inherent ability in reducing technical variations.
}
\examples{
\dontrun{
# Load a Seurat object with UCE embeddings
ifnb <- readRDS('ifnb_treatment.rds')
# subset the object to get CD8+ T cells
cd8t <- subset(ifnb, seurat_annotations == 'CD8 T')
# Use the function to measure CD8+ T cell polarization
# Cells from the control samples are used as unpolarized cells.
cd8t <- MeasurePolar(cd8t, celltype='CD8T',
unpolarized_cell=WhichCells(cd8t, expression = stim == 'CTRL'))
# Plot the CD8+ T cell polarization scores
FeaturePlot(cd8t, c('T8.a_score','T8.b_score','T8.c_score','T8.e_score','T8.f_score'))
# Plot the CD8+ T cell polarization P values
FeaturePlot(cd8t, c('T8.a_p','T8.b_p','T8.c_p','T8.e_p','T8.f_p'))
}

}
\references{
Cui A, Huang T, Li S, et al. Dictionary of immune responses to
cytokines at single-cell resolution. Nature. 2024;625(7994):377-384.
doi:10.1038/s41586-023-06816-9

Rosen Y, Roohani Y, et al. Universal Cell Embeddings: A Foundation Model for
Cell Biology. BioRxiv. 2023. doi: 10.1101/2023.11.28.568918
}
