% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polar_states}
\alias{polar_states}
\title{Polarization states}
\format{
A list with 66 elements of polarization states. The polarization states
are for 12 cell types: B, NK, CD8T, CD4T, Treg, Tgd, pDC, cDC1, cDC2, MigDC, LC,
Macro, Mono, Neu. Each element is a list containing 2 variables:
\describe{
\item{marker}{The 5 marker genes of the polarization state.}
\item{cytokine}{The driving cytokines of the polarization state.}
}
}
\usage{
polar_states
}
\description{
A list of immune cell polarization states with top marker genes and drivind cytokines.
}
\references{
Cui A, Huang T, Li S, et al. Dictionary of immune responses to
cytokines at single-cell resolution. Nature. 2024;625(7994):377-384.
doi:10.1038/s41586-023-06816-9
}
\keyword{datasets}
